Task_3_Resume_Screening_System/requirements.txt
langchain
langchain-core
langchain-huggingface
langsmith
python-dotenv
pydantic
